HL2 I/O board: send it the transverter's on-air transmit and receive frequencies instead of the I.F. (fixes #292 B)

+++ crates/sdroxide-hpsdr/src/ioboard.rs
@@ -173,7 +173,7 @@ pub(crate) struct IoBoard {
     queue: VecDeque<Op>,
     /// The transmit frequency the board has been told, so an unchanged dial is
     /// silent.
-    sent_hz: Option<u32>,
+    sent_hz: Option<u64>,
     /// The receive frequency *code* it has been told. Held as the code rather
     /// than the frequency because that is what the board acts on, and a code
     /// spans a whole band (see [`hertz_to_code`]).
@@ -205,12 +205,14 @@ impl IoBoard {
     /// this talks only when the frequency moves.
     ///
     /// `tx_freq_hz` is the frequency the radio would transmit on right now;
-    /// `rx_freq_hz` is where its receiver is tuned.
+    /// `rx_freq_hz` is where its receiver is tuned. Both are on the *air*: with
+    /// a transverter in front they are the transverter's band, not the I.F. the
+    /// radio itself is on, and a 3 cm dial does not fit in 32 bits.
     pub(crate) fn next_request(
         &mut self,
         now: Instant,
-        tx_freq_hz: u32,
-        rx_freq_hz: u32,
+        tx_freq_hz: u64,
+        rx_freq_hz: u64,
         mox: u8,
     ) -> Option<[u8; 5]> {
         // A request still waiting on its answer holds the bus: the protocol
@@ -237,7 +239,7 @@ impl IoBoard {
     }
 
     /// Queue whatever there is to say, if anything.
-    fn refill(&mut self, now: Instant, tx_freq_hz: u32, rx_freq_hz: u32) {
+    fn refill(&mut self, now: Instant, tx_freq_hz: u64, rx_freq_hz: u64) {
         if !self.queue.is_empty() {
             return;
         }
@@ -385,7 +387,7 @@ impl IoBoard {
 /// and preselector on, and deduplicating on the *code* rather than the frequency
 /// means a spun dial puts nothing at all on the I2C bus while a band change
 /// always does.
-fn hertz_to_code(hz: u32) -> u8 {
+fn hertz_to_code(hz: u64) -> u8 {
     if hz == 0 {
         return 0;
     }
@@ -398,10 +400,10 @@ fn hertz_to_code(hz: u32) -> u8 {
 /// The five register writes that hand `hz` to the board, **BYTE0 last** — that
 /// write is what makes the board act on the frequency, so the other four have to
 /// already be in place.
-fn freq_writes(hz: u32) -> [Op; 5] {
-    // A five-byte big-endian field. A 32-bit frequency in Hz never fills the
-    // top byte, which is there for transverted frequencies above 4.29 GHz.
-    let be = (hz as u64).to_be_bytes();
+fn freq_writes(hz: u64) -> [Op; 5] {
+    // A five-byte big-endian field; the top byte is there for transverted
+    // frequencies above 4.29 GHz.
+    let be = hz.to_be_bytes();
     [
         Op::write(REG_TX_FREQ_BYTE4, be[3]),
         Op::write(REG_TX_FREQ_BYTE4 + 1, be[4]),
@@ -417,7 +419,7 @@ mod tests {
 
     /// A receive frequency for the tests that are not about the receive code.
     /// Held fixed so it contributes exactly one write after the reset.
-    const RX_HZ: u32 = 14_074_000;
+    const RX_HZ: u64 = 14_074_000;
 
     /// Answer the request `board` just made, as the gateware would.
     fn ack(board: &mut IoBoard, now: Instant, data: [u8; 4]) {
